+++ plugins/woocommerce-admin/packages/components/src/search/autocompleters/index.js
@@ -5,6 +5,7 @@
 export { default as countries } from './countries';
 export { default as coupons } from './coupons';
 export { default as customers } from './customers';
+export { default as emails } from './emails';
 export { default as product } from './product';
 export { default as productCategory } from './product-cat';
 export { default as variations } from './variations';